feat: add some new SQL tables + add PHPDocBlocks to ConverterCsvToSql
