+++ data/favorites.csv
@@ -1,0 +1,27 @@
+user_id,selected_user_id
+1,2
+1,5
+1,9
+3,1
+3,2
+3,20
+7,10
+10,1
+10,7
+19,1
+19,2
+19,3
+19,4
+19,5
+19,6
+19,7
+19,8
+19,9
+19,10
+3,12
+3,13
+3,14
+5,2
+6,3
+7,4
+7,5
