+++ data/notices_types.csv
@@ -1,0 +1,7 @@
+type
+'new_response'
+'new_message'
+'task_reject'
+'task_start'
+'task_finish'
+
