+++ data/original/cities2.csv
@@ -1,4 +1,4 @@
-name
+﻿city,lat,long
 'Абаза',52.6517296,90.0885929
 'Абакан',53.7223661,91.4437792
 'Абдулино',53.6778096,53.6473115
@@ -78,6 +78,7 @@ name
 'Белинский',52.9641169,43.4185003
 'Белово',54.4221442,86.3036438
 'Белогорск',50.9212235,128.4738742
+'Белогорск',45.0571581,34.5998775
 'Белозерск',60.0307914,37.7891123
 'Белокуриха',51.9960534,84.9840343
 'Беломорск',64.5378468,34.7799717
@@ -89,6 +90,7 @@ name
 'Бердск',54.7582375,83.1072212
 'Березники',59.4079898,56.8040725
 'Березовский',55.6692853,86.2744636
+'Березовский',56.9096284,60.8180256
 'Беслан',43.1937634,44.5338746
 'Бийск',52.5393683,85.2138852
 'Бикин',46.8186244,134.2551046
@@ -98,6 +100,7 @@ name
 'Бирюсинск',55.9608611,97.8203624
 'Бирюч',50.6483899,38.4006203
 'Благовещенск',50.2905935,127.5272186
+'Благовещенск',55.0499592,55.955309
 'Благодарный',45.0990834,43.4307042
 'Бобров',51.0902048,40.0320014
 'Богданович',56.7764419,62.0463692
@@ -206,6 +209,7 @@ name
 'Гуково',48.0449422,39.9484635
 'Гулькевичи',45.3605769,40.6918484
 'Гурьевск',54.7705368,20.6039109
+'Гурьевск',54.2860005,85.947723
 'Гусев',54.5915477,22.1944114
 'Гусиноозерск',51.2865751,106.5230732
 'Давлеканово',54.2227617,55.0313419
@@ -260,6 +264,7 @@ name
 'Ефремов',53.1465725,38.0922412
 'Железноводск',44.1320689,43.0306475
 'Железногорск',56.2529346,93.532167
+'Железногорск',52.3378753,35.3516834
 'Жердевка',51.8421783,41.4617407
 'Жигулевск',53.4011545,49.494529
 'Жиздра',53.7522384,34.7386719
@@ -279,6 +284,7 @@ name
 'Заполярный',69.413232,30.7985332
 'Зарайск',54.7625993,38.8851485
 'Заречный',53.1961009,45.1691502
+'Заречный',56.8103345,61.3380592
 'Заринск',53.7063495,84.9314703
 'Звенигово',55.9739374,48.0170276
 'Звенигород',55.7296168,36.8552843
@@ -371,8 +377,10 @@ name
 'Киржач',56.1485263,38.8635035
 'Кириллов',59.8590353,38.3749854
 'Кириши',59.4742626,32.0625536
+'Киров',54.0790823,34.3077508
 'Киров',58.6035313,49.6679219
 'Кировград',57.4298451,60.0623588
+'Кировск',59.875371,30.9815205
 'Кировск',67.6150979,33.6637196
 'Кирс',59.3400586,52.2415895
 'Кирсанов',52.650594,42.7284872
@@ -419,15 +427,18 @@ name
 'Котовск',52.592393,41.5102486
 'Кохма',56.9325151,41.0933071
 'Красавино',60.9613991,46.481465
+'Красноармейск',51.0268595,45.7011086
 'Красноармейск',51.0236026,45.695141
 'Красновишерск',60.3902242,57.0537244
 'Красногорск',55.8318475,37.32942
 'Краснодар',45.040235,38.9760801
 'Краснозаводск',56.4408723,38.2320671
+'Краснознаменск',54.9452612,22.4928511
 'Краснознаменск',55.5978962,37.0395137
 'Краснокаменск',50.092858,118.0323407
 'Краснокамск',58.0820946,55.7479797
 'Красноперекопск',45.9537956,33.7921523
+'Краснослободск',48.7068777,44.5631799
 'Краснослободск',54.4248634,43.7845096
 'Краснотурьинск',59.7636934,60.1934451
 'Красноуральск',58.3486334,60.0409011
@@ -531,9 +542,11 @@ name
 'Миллерово',48.925821,40.3983302
 'Минусинск',53.7105509,91.6871866
 'Миньяр',55.0709601,57.5484435
+'Мирный',62.7644267,40.3359976
 'Мирный',62.5362511,113.9668146
 'Михайлов',54.2296493,39.0090857
 'Михайловка',50.0708749,43.2400847
+'Михайловск',56.4369736,59.113865
 'Михайловск',45.1296297,42.028787
 'Мичуринск',52.891294,40.510475
 'Могоча',53.7362465,119.7661232
@@ -581,6 +594,7 @@ name
 'Нижнекамск',55.6311963,51.814407
 'Нижнеудинск',54.8969105,99.0313481
 'Николаевск',50.0166952,45.4733204
+'Никольск',59.5350162,45.4575468
 'Никольск',53.7138397,46.079962
 'Никольское',59.7042633,30.7874879
 'Новоалександровск',45.4933561,41.2154482
@@ -624,6 +638,7 @@ name
 'Обь',54.9945882,82.6937925
 'Одинцово',55.6789292,37.2636389
 'Озерск',54.4084804,22.013508
+'Озерск',55.7631922,60.7076337
 'Озеры',54.8539944,38.55987
 'Октябрьск',53.1640886,48.6707463
 'Октябрьский',54.4815344,53.4656611
@@ -692,6 +707,7 @@ name
 'Пошехонье',58.506328,39.1210034
 'Правдинск',54.4430604,21.0083638
 'Приволжск',57.3807249,41.2809239
+'Приморск',54.7311185,19.9983108
 'Приморск',60.3660787,28.6135962
 'Приозерск',61.03313,30.15876
 'Прокопьевск',53.8604562,86.7183428
@@ -707,6 +723,7 @@ name
 'Пущино',54.832521,37.6210428
 'Пыталово',57.0637109,27.9234759
 'Пятигорск',44.0411669,43.0661091
+'Радужный',55.9961034,40.3323247
 'Радужный',62.1343067,77.4584346
 'Райчихинск',49.7941882,129.4112149
 'Раменское',55.549529,38.273015
@@ -784,6 +801,8 @@ name
 'Снежинск',56.085232,60.7324989
 'Снежногорск',69.1921469,33.2382821
 'Собинка',55.993798,40.0180034
+'Советск',55.0810474,21.8886272
+'Советск',57.5841434,48.9589228
 'Советск',53.9337099,37.6315517
 'Советский',61.3706695,63.566863
 'Сокол',59.4759156,40.1114414
@@ -899,6 +918,7 @@ name
 'Уяр',55.8131698,94.3282148
 'Фатеж',52.0918591,35.8538912
 'Феодосия',45.031878,35.3824772
+'Фокино',53.4554552,34.4158537
 'Фокино',42.9707163,132.4110209
 'Фролово',49.7647809,43.6648936
 'Фрязино',55.9590154,38.0409171
@@ -984,6 +1004,7 @@ name
 'Ядрин',55.9406826,46.2020765
 'Якутск',62.0280273,129.7325717
 'Ялта',44.416366,34.04455
+'Ялта',44.4951948,34.166386
 'Ялуторовск',56.649167,66.297034
 'Янаул',56.2650336,54.9298314
 'Яранск',57.304195,47.8479185
