+++ data/original/notices_types.csv
@@ -1,0 +1,9 @@
+name,icon_link
+'Курьерские услуги','translation'
+'Уборка','clean'
+'Переезды','cargo'
+'Компьютерная помощь','neo'
+'Ремонт квартирный','flat'
+'Ремонт техники','repair'
+'Красота','beauty'
+'Фото','photo'
