+++ data/original/profiles.csv
@@ -1,16 +1,16 @@
 address,bd,about,phone,skype
 38737 Moose Avenue,1989-11-11,"In est risus, auctor sed, tristique in, tempus sit amet, sem. Fusce consequat.",
-64574473047,high-level
+64574473047,high#level
 738 Hagan Lane,1989-03-05,Pellentesque ultrices mattis odio.,75531015353,mobile
-758 Old Shore Parkway,1989-12-30,Morbi a ipsum. Integer a nibh. In quis justo.,16371407508,Re-engineered
-11 Dovetail Junction,0629-03-03,Suspendisse potenti.,21468788926,Grass-roots
-050 Bowman Alley,1989-04-08,"Morbi quis tortor id nulla ultrices aliquet. Maecenas leo odio, condimentum id, luctus nec, molestie sed, justo.",62931646367,fault-tolerant
-5 Iowa Avenue,1989-04-18,"Maecenas tristique, est et tempus semper, est quam pharetra magna, ac consequat metus sapien ut nunc. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia Curae; Mauris viverra diam vitae quam.",63271348718,Team-oriented
+758 Old Shore Parkway,1989-12-30,Morbi a ipsum. Integer a nibh. In quis justo.,16371407508,Re#engineered
+11 Dovetail Junction,0629-03-03,Suspendisse potenti.,21468788926,Grass#roots
+050 Bowman Alley,1989-04-08,"Morbi quis tortor id nulla ultrices aliquet. Maecenas leo odio, condimentum id, luctus nec, molestie sed, justo.",62931646367,fault#tolerant
+5 Iowa Avenue,1989-04-18,"Maecenas tristique, est et tempus semper, est quam pharetra magna, ac consequat metus sapien ut nunc. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia Curae; Mauris viverra diam vitae quam.",63271348718,Team#oriented
 87119 Northland Hill,1989-03-20,Sed vel enim sit amet nunc viverra dapibus. Nulla suscipit ligula in lacus.,41056175169,portal
 6823 Lillian Point,1989-12-13,"Pellentesque viverra pede ac diam. Cras pellentesque volutpat dui. Maecenas tristique, est et tempus semper, est quam pharetra magna, ac consequat metus sapien ut nunc.",72882384431,intermediate
 43 Marquette Plaza,1989-01-14,Morbi ut odio.,69043821394,local area network
-5303 Village Green Hill,1989-02-03,"Suspendisse ornare consequat lectus. In est risus, auctor sed, tristique in, tempus sit amet, sem. Fusce consequat.",28396220507,upward-trending
-67399 Reindahl Place,1989-05-23,Phasellus sit amet erat. Nulla tempus. Vivamus in felis eu sapien cursus vestibulum.,83344513307,grid-enabled
+5303 Village Green Hill,1989-02-03,"Suspendisse ornare consequat lectus. In est risus, auctor sed, tristique in, tempus sit amet, sem. Fusce consequat.",28396220507,upward#trending
+67399 Reindahl Place,1989-05-23,Phasellus sit amet erat. Nulla tempus. Vivamus in felis eu sapien cursus vestibulum.,83344513307,grid#enabled
 45 Twin Pines Hill,1989-07-06,"Vivamus metus arcu, adipiscing molestie, hendrerit at, vulputate vitae, nisl. Aenean lectus. Pellentesque eget nunc.",64890419671,background
 46 Sheridan Place,1903-04-16,Quisque ut erat. Curabitur gravida nisi at nibh. In hac habitasse platea dictumst.,23005580487,challenge
 73 Kedzie Terrace,1989-11-07,Pellentesque at nulla. Suspendisse potenti. Cras in purus eu magna vulputate luctus.,27052074771,coherent
