+++ data/original/users.csv
@@ -1,0 +1,21 @@
+email,name,password,dt_add
+kbuttress0@1und1.de,Karrie Buttress,JcfoKBYAB4k,2019-08-10
+baymer1@hp.com,Bob Aymer,ZEE54kg,2018-12-21
+zboulding2@macromedia.com,Zilvia Boulding,VJyMV1Zat,2019-07-25
+emollon3@bloglovin.com,Emalee Mollon,XUIeJ693h,2018-11-13
+mmulberry4@cmu.edu,Maria Mulberry,oWspnl,2019-07-20
+lby5@mozilla.com,Levey By,GdtcUU,2019-02-12
+beates6@last.fm,Baron Eates,UQw6VeA,2019-05-03
+tvink7@fotki.com,Trip Vink,49znXd7haFGz,2019-01-13
+bterbeck8@about.me,Boonie Terbeck,unCjJTF7sjs,2019-09-15
+atraviss9@auda.org.au,Alonzo Traviss,dLuVMAg,2018-12-19
+nwitteringa@google.com.br,Natassia Wittering,tQlUG4n,2019-03-24
+fbrookeb@nba.com,Felice Brooke,s9y9Mcfgy1g,2019-09-27
+cviccaryc@amazon.co.uk,Carlen Viccary,9qd747vh,2018-12-06
+hgethingsd@sogou.com,Hendrik Gethings,zzN5c4,2018-11-18
+dgirodiase@stanford.edu,Dunc Girodias,j9QW6GQI,2018-10-14
+btanmanf@smh.com.au,Bibbie Tanman,1aukKNEIneq,2019-05-03
+bbartolettig@simplemachines.org,Barnabas Bartoletti,3chTNtqhoo,2018-12-25
+nculliph@fc2.com,Nixie Cullip,2UdKIR2f,2019-04-07
+mpimblotti@xing.com,Matilde Pimblott,nGZ8disdg,2019-07-18
+askurrayj@un.org,Al Skurray,bL9tAf,2018-11-25
