+++ sql/categories_2.sql
