+++ sql/notices_types_1.sql
@@ -1,0 +1,6 @@
+INSERT INTO `tforce`.`notices_types` (`type`) VALUES 
+ ('new_response'), 
+ ('new_message'), 
+ ('task_reject'), 
+ ('task_start'), 
+ ('task_finish');
