+++ sql/users_4.sql
@@ -1,0 +1,21 @@
+INSERT INTO `tforce`.`users` (`email`, `phio`, `password_hash`, `created_at`, `city_id`, `open_profile`, `open_contacts`, `last_activity`, `biography`, `birthday`, `phone`, `skype`, `rate`, `telegram`, `views_number`, `fail_count`) VALUES 
+ ('kbuttress0@1und1.de', 'Karrie Buttress', 'JcfoKBYAB4k', '2019-08-10 01:02:03', 5, 1, 0, '2020-01-01  02:00:00', 'biography 111', '1989-11-11', '64574473047', 'skype10', 4, 'telegram111', 10, 0), 
+ ('baymer1@hp.com', 'Bob Aymer', 'ZEE54kg', '2018-12-21 04:05:06', 6, 1, 1, '2019-02-01  03:20:01', 'biography 222', '1989-03-05', '75531015353', 'skype11', 3, 'telegram222', 20, 0), 
+ ('zboulding2@macromedia.com', 'Zilvia Boulding', 'VJyMV1Zat', '2019-07-25 07:08:09', 100, 0, 0, '2019-10-15  04:30:02', 'biography 333', '1989-12-30', '16371407508', NULL, 2, 'telegram333', 0, 2), 
+ ('emollon3@bloglovin.com', 'Emalee Mollon', 'XUIeJ693h', '2018-11-13 10:11:12', 20, 0, 0, '2019-10-05  05:40:03', 'biography 444', '1989-03-03', '21468788926', 'skype12', 1.2, 'telegram444', 30, 3), 
+ ('mmulberry4@cmu.edu', 'Maria Mulberry', 'oWspnl', '2019-07-20 13:14:15', 5, 1, 1, '2019-08-04  06:50:04', 'biography 555', '1989-04-08', '62931646367', 'skype13', 4, NULL, 40, 4), 
+ ('lby5@mozilla.com', 'Levey By', 'GdtcUU', '2019-02-12 16:17:18', 20, 1, 0, '2019-08-01  07:10:07', 'biography 666', '1989-04-18', '63271348718', 'skype14', 4, 'telegram555', 50, 5), 
+ ('beates6@last.fm', 'Baron Eates', 'UQw6VeA', '2019-05-03 09:29:28', 21, 0, 0, '2019-06-25  08:20:08', 'biography 777', '1989-03-20', '41056175169', 'skype15', 4.5, 'telegram666', 45, 6), 
+ ('tvink7@fotki.com', 'Trip Vink', '49znXd7haFGz', '2019-01-13 08:27:26', 100, 0, 1, '2019-01-24  09:34:22', 'biography 888', '1989-12-13', '72882384431', 'skype16', 5, 'telegram777', 55, 7), 
+ ('bterbeck8@about.me', 'Boonie Terbeck', 'unCjJTF7sjs', '2019-09-15 07:25:24', 5, 0, 0, '2019-12-23  10:44:22', 'biography 999', '1989-01-14', '69043821394', 'skype17', 3, 'telegram888', 0, 8), 
+ ('atraviss9@auda.org.au', 'Alonzo Traviss', 'dLuVMAg', '2018-12-19 06:23:22', 20, 1, 1, '2019-11-22  11:55:33', 'biography 101', '1989-02-03', '28396220507', 'skype18', 3, NULL, 65, 7), 
+ ('nwitteringa@google.com.br', 'Natassia Wittering', 'tQlUG4n', '2019-03-24 05:21:20', 6, 1, 0, '2019-10-21  12:22:44', 'biography 121', '1989-05-23', '83344513307', 'skype19', 3.1, NULL, 75, 2), 
+ ('fbrookeb@nba.com', 'Felice Brooke', 's9y9Mcfgy1g', '2019-09-27 04:19:18', 6, 0, 0, '2019-10-10  13:11:55', 'biography 131', '1989-07-06', '64890419671', 'skype310', 3.1, 'telegram999', 12, 3), 
+ ('cviccaryc@amazon.co.uk', 'Carlen Viccary', '9qd747vh', '2018-12-06 03:17:16', 6, 1, 1, '2019-01-09  14:23:5', 'biography 141', '1903-04-16', '23005580487', 'skype320', 0, 'telegram122', 13, 2), 
+ ('hgethingsd@sogou.com', 'Hendrik Gethings', 'zzN5c4', '2018-11-18 02:15:14', 6, 0, 0, '2019-02-08  15:33:10', 'biography 151', '1989-11-07', '27052074771', NULL, NULL, 'telegram123', 14, 0), 
+ ('dgirodiase@stanford.edu', 'Dunc Girodias', 'j9QW6GQI', '2018-10-14 01:13:12', 200, 1, 1, '2019-03-07  16:44:15', 'biography 161', '1989-02-13', '14800371520', NULL, NULL, 'telegram124', 15, 2), 
+ ('btanmanf@smh.com.au', 'Bibbie Tanman', '1aukKNEIneq', '2019-05-03 00:11:10', 300, 0, 0, '2019-06-06  17:55:20', 'biography 171', '1989-07-07', '75569924500', NULL, NULL, 'telegram125', 16, 2), 
+ ('bbartolettig@simplemachines.org', 'Barnabas Bartoletti', '3chTNtqhoo', '2018-12-25 00:00:09', 401, 1, 0, '2020-01-05  18:22:25', 'biography 181', '1989-09-29', '37349256497', NULL, NULL, NULL, 17, 3), 
+ ('nculliph@fc2.com', 'Nixie Cullip', '2UdKIR2f', '2019-04-07 07:00:08', 100, 1, 0, '2020-01-04  19:13:30', 'biography 191', '1989-03-19', '12403580562', 'skype1444', 4.1, 'telegram126', 17, 3), 
+ ('mpimblotti@xing.com', 'Matilde Pimblott', 'nGZ8disdg', '2019-07-18 06:00:05', 100, 0, 1, '2020-01-03  22:42:35', 'biography 231', '1989-09-29', '40139478003', NULL, 4.1, 'telegram127', 17, 0), 
+ ('askurrayj@un.org', 'Al Skurray', 'bL9tAf', '2018-11-25 04:00:03', 100, 1, 0, '2020-01-02  23:02:40', 'biography 232', '1989-12-21', '76657531985', NULL, 4, 'telegram7897', 17, 1);
